% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertFalse}
\alias{assertFalse}
\title{Assert that the value evaluates to false}
\usage{
assertFalse(value)
}
\arguments{
\item{value}{a vector with any length.}
}
\description{
Asserts that the given \code{value} evaluates to \code{FALSE}. If the value
evaluates does not evaluate to \code{FALSE}, then an error condition is
raised.
}
\details{
The \code{\link{identical}} function is used to check this condition,
so this function will also raise an error condition if \code{value} evaluates
to \code{0} or if the result of the evaluation has any attributes.
}
\examples{
\dontrun{
assertFalse(is.character(1L))
## If you use an R version greater than 3.5.0, you can also use:
assertThat(is.numeric("a"), isFALSE)
}
}
\seealso{
\itemize{
\item \code{\link{assertThat}}
\item \code{\link{assertTrue}}
\item \code{\link{isFalse}}
}
}
