% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertTrue}
\alias{assertTrue}
\title{Assert that the value evaluates to true}
\usage{
assertTrue(value)
}
\arguments{
\item{value}{a vector with any length.}
}
\description{
Assert that the value evaluates to true
}
\examples{
\dontrun{
assertTrue(is.numeric(2019))
}
}
\seealso{
\itemize{
\item \code{\link{assertThat}}
\item \code{\link{assertFalse}}
\item \code{\link{isTrue}}
}
}
