% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{closeTo}
\alias{closeTo}
\title{Returns a function that checks whether its \emph{argument} is close to an
\emph{expected} result within the given tolerance}
\usage{
closeTo(expected, delta)
}
\arguments{
\item{expected}{a numeric vector.}

\item{delta}{a numeric vector of length one that defines the maximum allowed
difference.}
}
\description{
Returns a function that checks whether its \emph{argument} is close to an
\emph{expected} result within the given tolerance
}
\details{
The \code{closeTo()} matcher function can only be used for numeric arguments.
}
\examples{
\dontrun{
assertThat(-0.50557992900139, closeTo(-0.50557, delta = 1e4))
}
}
\seealso{
\itemize{
\item \code{\link{identicalTo}}
\item \code{\link{equalTo}}
}
}
