% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest-package.R
\docType{package}
\name{hamcrest-package}
\alias{hamcrest}
\alias{hamcrest-package}
\title{Unit tests using the Hamcrest framework}
\description{
The Hamcrest framework for writing unit tests is based on an assertion
function and matcher objects. A matcher object implements a single rule that
is used by the assertion function to perform the test. This package provides
a small set of matcher objects in the form of functions. The functionality
provided by this set of matcher objects is identical to those implemented in
Renjin, the R interpreter for the Java Virtual Machine. The goal of this
package is therefore to be able to run Renjin's unit tests written in R
through the GNU R interpreter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedatadriven/hamcrest}
  \item Report bugs at \url{https://github.com/bedatadriven/hamcrest/issues}
}

}
\author{
\strong{Maintainer}: Maarten-Jan Kallen \email{mj@bedatadriven.com} (0000-0002-1839-5333)

Authors:
\itemize{
  \item Alexander Bertram
  \item Parham Solaimani (0000-0001-5292-1379)
  \item Metin Yazici
}

Other contributors:
\itemize{
  \item BeDataDriven [copyright holder, funder]
}

}
\keyword{internal}
