% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{Assert that the value is identical to an expected value}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function.}

\item{tol}{numeric tolerance.}
}
\description{
Assert that the value is identical to an expected value
}
\examples{
\dontrun{
assertThat(floor(-1.5), identicalTo(-2))
}
}
\seealso{
\itemize{
\item \code{\link{closeTo}}
\item \code{\link{equalTo}}
}
}
