% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{isFalse}
\alias{isFalse}
\title{Assert that the actual value is false}
\usage{
isFalse()
}
\description{
Assert that the actual value is false
}
\examples{
\dontrun{
assertThat(is.character(seq(10)), isFalse())
}
}
\seealso{
\itemize{
\item \code{\link{assertTrue}}
\item \code{\link{assertFalse}}
\item \code{\link{isTrue}}
}
}
