% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{isTrue}
\alias{isTrue}
\title{Checks if the expected value is true}
\usage{
isTrue()
}
\description{
Checks if the expected value is true
}
\examples{
\dontrun{
assertThat(is.integer(1L), isTrue())
}
}
\seealso{
\itemize{
\item \code{\link{assertTrue}}
\item \code{\link{assertFalse}}
\item \code{\link{isFalse}}
}
}
