% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{testHamcrest}
\alias{testHamcrest}
\title{A testing suite for hamcrest tests}
\usage{
testHamcrest(note, ...)
}
\arguments{
\item{note}{an explanatory text for the tests.}

\item{...}{hamcrest assertions.}
}
\description{
The tests put in this suite will be evaluated in a local environment. It is
suitable to run hamcrest tests without being bothered by the global
namespace.
}
