% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{throw_assert_error}
\alias{throw_assert_error}
\title{Throw assert error used when writing custom assertions}
\usage{
throw_assert_error(value)
}
\arguments{
\item{value}{a vector with any length.}
}
\description{
Throw assert error used when writing custom assertions
}
