% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertFalse}
\alias{assertFalse}
\title{Assert that the value evaluates to false}
\usage{
assertFalse(value)
}
\arguments{
\item{value}{a vector with any length.}
}
\description{
Assert that the value evaluates to false
}
\examples{
\dontrun{
assertFalse(is.character(1L))
}
}
\seealso{
\code{\link{assertThat}}\cr
\code{\link{assertTrue}}\cr
\code{\link{isFalse}}
}
