% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertFalse}
\alias{assertFalse}
\title{Is the matched assertion false?}
\usage{
assertFalse(value)
}
\arguments{
\item{value}{a vector.}
}
\description{
Is the matched assertion false?
}
\examples{
\dontrun{
assertFalse(is.character(1L))
}
}
\seealso{
\code{\link{isFalse}}
}
