% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertTrue}
\alias{assertTrue}
\title{assertTrue}
\usage{
assertTrue(value)
}
\arguments{
\item{value}{a vector}
}
\description{
assertTrue
}
\examples{
\dontrun{
assertTrue(is.numeric(2019))
}
}
\seealso{
isTrue
}
