% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertTrue}
\alias{assertTrue}
\title{Assert that the value evaluates to true}
\usage{
assertTrue(value)
}
\arguments{
\item{value}{a vector with any length.}
}
\description{
Asserts that the given \code{value} evaluates to \code{TRUE}. If the value
evaluates does not evaluate to \code{TRUE}, then an error condition is
raised.
}
\details{
The \code{\link{identical}} function is used to check this condition,
so this function will also raise an error condition if \code{value} evaluates
to \code{0} or if the result of the evaluation has any attributes.
}
\examples{
\dontrun{
assertTrue(is.numeric(2019))
assertThat(is.character("orange"), isTRUE)
}
}
\seealso{
\itemize{
\item \code{\link{assertThat}}
\item \code{\link{assertFalse}}
\item \code{\link{isTrue}}
}
}
