% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{assertTrue}
\alias{assertTrue}
\title{Is the matched assertion true?}
\usage{
assertTrue(value)
}
\arguments{
\item{value}{a vector.}
}
\description{
Is the matched assertion true?
}
\examples{
\dontrun{
assertTrue(is.numeric(2019))
}
}
\seealso{
\code{\link{isTrue}}
}
