% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{closeTo}
\alias{closeTo}
\title{How much do actual value(s) close to expected value(s)?}
\usage{
closeTo(expected, delta)
}
\arguments{
\item{expected}{a numeric vector.}

\item{delta}{a numeric vector of length one that defines the maximum allowed
difference}
}
\description{
How much do actual value(s) close to expected value(s)?
}
\examples{
\dontrun{
assertThat(-0.50557992900139, closeTo(-0.50557, delta = 1e4))
}
}
\seealso{
\code{\link{identicalTo}},\code{\link{equalTo}}
}
