% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{closeTo}
\alias{closeTo}
\title{Checks if the expected value is close to a value within some tolerance level}
\usage{
closeTo(expected, delta)
}
\arguments{
\item{expected}{a numeric vector.}

\item{delta}{a numeric vector of length one that defines the maximum allowed
difference.}
}
\description{
A \emph{matcher} function which will return \code{TRUE} if its first argument
is a numeric vector of length 1 whose absolute difference with expected is
greater than or equal to the value of \code{delta}.
}
\details{
The \code{closeTo()} matcher function can only be used for numeric arguments.
}
\examples{
\dontrun{
assertThat(-0.50557992900139, closeTo(-0.50557, delta = 1e4))
}
}
\seealso{
\itemize{
\item \code{\link{identicalTo}}
\item \code{\link{equalTo}}
}
}
