% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{closeTo}
\alias{closeTo}
\title{closeTo}
\usage{
closeTo(expected, delta)
}
\arguments{
\item{expected}{object passed to the matcher function. For \code{closeTo()}
this argument must be a numeric vector.}

\item{delta}{a numeric vector of length one that defines the maximum allowed
difference}
}
\description{
closeTo
}
