% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{deparsesTo}
\alias{deparsesTo}
\title{Assert that the actual value deparses to an expected value}
\usage{
deparsesTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function.}
}
\description{
Deparsing transforms unevaluated expressions into character vectors.
}
\examples{
\dontrun{
assertThat(unlist(quote(sin(3.14)), recursive = FALSE), deparsesTo("sin(3.14)"))
}
}
