% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{emitsWarning}
\alias{emitsWarning}
\title{Returns a function that checks whether its \emph{argument} emits warning(s)}
\usage{
emitsWarning()
}
\description{
Returns a function that checks whether its \emph{argument} emits warning(s)
}
\examples{
\dontrun{
assertThat(any(range(2.0,3.0)), emitsWarning())
}
}
\seealso{
\itemize{
\item \code{\link{throwsError}}
}
}
