% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{emitsWarning}
\alias{emitsWarning}
\title{Assert that the actual value emits warning}
\usage{
emitsWarning()
}
\description{
A \emph{matcher} function which will return \code{TRUE} if the actual value
is identical to the expected value.
}
\examples{
\dontrun{
assertThat(any(range(2.0,3.0)), emitsWarning())
}
}
\seealso{
\itemize{
\item \code{\link{throwsError}}
}
}
