% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{equalTo}
\alias{equalTo}
\title{Checks if the expected value is equal to a value}
\usage{
equalTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function.}
}
\description{
A \emph{matcher} function which will return \code{TRUE} if the actual value
is equal to the expected value.
}
\examples{
\dontrun{
assertThat(qnorm(0, 0, 1, TRUE, FALSE), equalTo(-Inf))
assertThat(length(rnorm(10)), equalTo(10))
}
}
\seealso{
\itemize{
\item \code{\link{identicalTo}}
\item \code{\link{closeTo}}
}
}
