% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{equalTo}
\alias{equalTo}
\title{equalTo}
\usage{
equalTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function. For \code{closeTo()}
this argument must be a numeric vector.}
}
\description{
equalTo
}
