% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{equalTo}
\alias{equalTo}
\title{equalTo}
\usage{
equalTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function.}
}
\description{
equalTo
}
\examples{
\dontrun{
assertThat(qnorm(0, 0, 1, TRUE, FALSE), equalTo(-Inf))
}
}
