% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{equalTo}
\alias{equalTo}
\title{Assert that the value is equal to an expected value}
\usage{
equalTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function.}
}
\description{
Assert that the value is equal to an expected value
}
\examples{
\dontrun{
assertThat(qnorm(0, 0, 1, TRUE, FALSE), equalTo(-Inf))
}
}
\seealso{
\itemize{
\item \code{\link{identicalTo}}
\item \code{\link{closeTo}}
}
}
