% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{equalTo}
\alias{equalTo}
\title{Is actual value equal to expected value?}
\usage{
equalTo(expected)
}
\arguments{
\item{expected}{object passed to the matcher function.}
}
\description{
Is actual value equal to expected value?
}
\examples{
\dontrun{
assertThat(qnorm(0, 0, 1, TRUE, FALSE), equalTo(-Inf))
}
\\code{\\link{identicalTo}},\\code{\\link{closeTo}}
}
