% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{Is actual value identical to expected value?}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function.}

\item{tol}{numeric tolerance.}
}
\description{
Is actual value identical to expected value?
}
\examples{
\dontrun{
assertThat(floor(-1.5), identicalTo(-2))
}
}
\seealso{
\code{\link{closeTo}},\code{\link{equalTo}}
}
