% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{Checks if the expected value is identical to a value}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function.}

\item{tol}{numeric tolerance.}
}
\description{
A \emph{matcher} function which will return \code{TRUE} if the actual value
is identical to the expected value.
}
\examples{
\dontrun{
assertThat(floor(-1.5), identicalTo(-2))
}
}
\seealso{
\itemize{
\item \code{\link{closeTo}}
\item \code{\link{equalTo}}
}
}
