% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{identicalTo}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function. For \code{closeTo()}
this argument must be a numeric vector.}

\item{tol}{numeric tolerance.}
}
\description{
identicalTo
}
