% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{Returns a function that checks whether its \emph{argument} is identical to an
\emph{expected} result}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function.}

\item{tol}{numeric tolerance. Should note that \code{tol} is only used if the
expected value is a finite floating point number. If the expected value is
another type, such as an integer, an exact match is expected.}
}
\description{
Returns a function that checks whether its \emph{argument} is identical to an
\emph{expected} result
}
\examples{
\dontrun{
assertThat(floor(-1.5), identicalTo(-2))
}
}
\seealso{
\itemize{
\item \code{\link{closeTo}}
\item \code{\link{equalTo}}
}
}
