% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{identicalTo}
\alias{identicalTo}
\title{identicalTo}
\usage{
identicalTo(expected, tol = NULL)
}
\arguments{
\item{expected}{object passed to the matcher function.}

\item{tol}{numeric tolerance.}
}
\description{
identicalTo
}
\examples{
\dontrun{
assertThat(floor(-1.5), identicalTo(-2))
}
}
