% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{isFalse}
\alias{isFalse}
\title{Returns a function that checks whether its \emph{argument} is 'FALSE'}
\usage{
isFalse()
}
\description{
Returns a function that checks whether its \emph{argument} is 'FALSE'
}
\examples{
\dontrun{
assertThat(is.character(seq(10)), isFalse())
}
}
\seealso{
\itemize{
\item \code{\link{assertTrue}}
\item \code{\link{assertFalse}}
\item \code{\link{isTrue}}
}
}
