% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamcrest.R
\name{throwsError}
\alias{throwsError}
\title{throwsError}
\usage{
throwsError()
}
\description{
throwsError
}
\examples{
\dontrun{
assertThat(log("a"), throwsError())
}
}
\seealso{
\itemize{
\item \code{\link{emitsWarning}}
}
}
